# asset-service/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Asset Management Service API
  description: |
    Asset Management Service API
    
    Comprehensive asset lifecycle management including:
    - Asset CRUD operations
    - Master data management (Categories, Makes, Models, Vendors, etc.)
    - User-Asset linking
    - Compliance and validation
    - Document management
    - Warranty and AMC management
    - Audit logging
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8082
    description: Local development server
  - url: https://api.example.com/asset-service
    description: Production server
...
paths:
...
  /api/asset/v1/assets/{id}:
    get:
      tags:
        - Assets
      summary: Get asset by ID
      description: Get asset details by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Asset found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
        '404':
          description: Asset not found

    put:
      tags:
        - Assets
      summary: Update asset
      description: Update an existing asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
      responses:
        '200':
          description: Asset updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
        '404':
          description: Asset not found

    delete:
      tags:
        - Assets
      summary: Delete asset
      description: Soft delete an asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Asset deleted successfully
        '404':
          description: Asset not found

  # ============================================================
  # CATEGORY ENDPOINTS
  # ============================================================
...
  /api/asset/v1/warranty:
    post:
      tags:
        - Warranty
      summary: Create warranty
      description: Create warranty information for an asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assetId
              properties:
                assetId:
                  type: integer
                  format: int64
                warrantyStartDate:
                  type: string
                  format: date
                warrantyEndDate:
                  type: string
                  format: date
                warrantyProvider:
                  type: string
                warrantyNumber:
                  type: string
      responses:
        '200':
          description: Warranty created
  /api/asset/v1/warranty/{id}:
    get:
      tags:
        - Warranty
      summary: Get warranty by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Warranty found

    put:
      tags:
        - Warranty
      summary: Update warranty
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Warranty updated

  # ============================================================
  # AMC ENDPOINTS
  # ============================================================
...
  /api/asset/v1/audit:
    get:
      tags:
        - Audit
      summary: Get all audit logs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /api/asset/v1/audit/username/{username}:
    get:
      tags:
        - Audit
      summary: Get audit logs by username
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of audit logs
  /api/asset/v1/audit/date-range:
    get:
      tags:
        - Audit
      summary: Get audit logs by date range
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of audit logs
  /api/asset/v1/audit/statistics:
    get:
      tags:
        - Audit
      summary: Get audit statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit statistics
          content:
            application/json:
              schema:
                type: object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseWrapper:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    AssetRequest:
      type: object
      required:
        - categoryId
        - subCategoryId
        - makeId
        - modelId
        - assetNameUdv
      properties:
        categoryId:
          type: integer
          format: int64
        subCategoryId:
          type: integer
          format: int64
        makeId:
          type: integer
          format: int64
        modelId:
          type: integer
          format: int64
        assetNameUdv:
          type: string
        userId:
          type: integer
          format: int64
        username:
          type: string
        projectType:
          type: string
...